+++ tokio/src/net/tcp/stream.rs
@@ -1,7 +1,6 @@
 use crate::future::poll_fn;
-use crate::io::{AsyncRead, AsyncWrite};
+use crate::io::{AsyncRead, AsyncWrite, PollEvented};
 use crate::net::tcp::split::{split, ReadHalf, WriteHalf};
-use crate::net::util::PollEvented;
 use crate::net::ToSocketAddrs;
 
 use bytes::{Buf, BufMut};
