+++ tokio/src/net/udp/socket.rs
@@ -1,6 +1,6 @@
 use crate::future::poll_fn;
+use crate::io::PollEvented;
 use crate::net::udp::split::{split, UdpSocketRecvHalf, UdpSocketSendHalf};
-use crate::net::util::PollEvented;
 use crate::net::ToSocketAddrs;
 
 use std::convert::TryFrom;
