--- tokio/src/net/util/mod.rs
@@ -1,4 +1,0 @@
-//! Utilities for implementing networking types.
-
-mod poll_evented;
-pub use self::poll_evented::PollEvented;
